# apps/api/src/server.ts
/**
 * @emailed/api — Main Server Entry Point
 *
 * Creates the Hono application, registers all routes, applies middleware,
 * starts listening, and handles graceful shutdown.
 *
 * This file is the production entry point. The `index.ts` file re-exports
 * for Bun's built-in HTTP server, but this module can also be used
 * standalone or in tests.
 */

import { Hono } from "hono";
import { cors } from "hono/cors";
import { requestId } from "hono/request-id";
import { logger } from "hono/logger";
import { timing } from "hono/timing";
import { secureHeaders } from "hono/secure-headers";

import { authMiddleware } from "./middleware/auth.js";
import {
  globalIpRateLimit,
  authRateLimit,
  sendRateLimit,
  readRateLimit,
  writeRateLimit,
  webhookRateLimit,
  searchRateLimit,
  closeRateLimitRedis,
} from "./middleware/rate-limit.js";
import { messages } from "./routes/messages.js";
import { domains } from "./routes/domains.js";
import { webhooks } from "./routes/webhooks.js";
import { analytics } from "./routes/analytics.js";
import { suppressions } from "./routes/suppressions.js";
import { tracking } from "./routes/tracking.js";
import { apiKeysRouter } from "./routes/api-keys.js";
import { account } from "./routes/account.js";
import { auth } from "./routes/auth.js";
import { health } from "./routes/health.js";
import { admin } from "./routes/admin.js";
import { billing } from "./routes/billing.js";
import { templatesRouter } from "./routes/templates.js";
import { voice } from "./routes/voice.js";
import { closeConnection } from "@emailed/db";
import { closeSendQueue } from "./lib/queue.js";
import { startWebhookWorker, stopWebhookWorker } from "./lib/webhook-dispatcher.js";
import { initSearchIndex, initTelemetry, shutdownTelemetry, telemetryMiddleware } from "@emailed/shared";

// ─── Create the Hono app ───────────────────────────────────────────────────

const app = new Hono();

// ─── Global Middleware ──────────────────────────────────────────────────────

// OpenTelemetry tracing and metrics
app.use("*", telemetryMiddleware());

// Request ID for distributed tracing
app.use("*", requestId());

// Structured request logging
app.use("*", logger());

// Server-Timing headers for performance debugging
app.use("*", timing());

// Security headers (CSP, X-Frame-Options, etc.)
app.use("*", secureHeaders());

// Global IP-based rate limit (DDoS baseline: 1000 req/min per IP)
app.use("*", globalIpRateLimit);

// CORS
app.use(
  "*",
  cors({
    origin: (origin) => origin,
    allowMethods: ["GET", "POST", "PUT", "PATCH", "DELETE", "OPTIONS"],
    allowHeaders: [
      "Authorization",
      "Content-Type",
      "X-API-Key",
      "X-Request-Id",
    ],
    exposeHeaders: [
      "X-Request-Id",
      "X-RateLimit-Limit",
      "X-RateLimit-Remaining",
      "X-RateLimit-Reset",
      "Retry-After",
    ],
    maxAge: 86400,
    credentials: true,
  }),
);

// ─── Health check (no auth) ────────────────────────────────────────────────

// Simple liveness probe (always returns 200)
app.get("/health", (c) => {
  return c.json({
    status: "ok",
    service: "emailed-api",
    version: process.env["SERVICE_VERSION"] ?? "0.1.0",
    timestamp: new Date().toISOString(),
  });
});

// Deep health check with dependency verification (also no auth)
app.route("/v1/health", health);

// Auth endpoints: strict IP rate limiting (10 req/min), no API key auth
app.use("/v1/auth/*", authRateLimit);
app.route("/v1/auth", auth);

// Tracking endpoints (no auth — embedded in emails)
app.route("/t", tracking);

// ─── Authenticated routes ──────────────────────────────────────────────────

// ─── Per-category rate limits on authenticated routes ─────────────────────
// Send endpoint: 100 req/min per API key
app.use("/v1/messages/send", authMiddleware, sendRateLimit);
// Search endpoint: 60 req/min per API key
app.use("/v1/messages/search", authMiddleware, searchRateLimit);
// Read messages: 600 req/min per API key
app.use("/v1/messages/*", authMiddleware, readRateLimit);
// Domains: write-level limits (200 req/min)
app.use("/v1/domains/*", authMiddleware, writeRateLimit);
// Webhooks: write-level limits (200 req/min)
app.use("/v1/webhooks/*", authMiddleware, writeRateLimit);
// Analytics: read-level limits (600 req/min)
app.use("/v1/analytics/*", authMiddleware, readRateLimit);
// Suppressions: write-level limits (200 req/min)
app.use("/v1/suppressions/*", authMiddleware, writeRateLimit);
// API keys management: write-level limits (200 req/min)
app.use("/v1/api-keys/*", authMiddleware, writeRateLimit);
// Account management: write-level limits (200 req/min)
app.use("/v1/account/*", authMiddleware, writeRateLimit);
// Billing authenticated endpoints: write-level limits (200 req/min)
app.use("/v1/billing/checkout", authMiddleware, writeRateLimit);
app.use("/v1/billing/portal", authMiddleware, writeRateLimit);
app.use("/v1/billing/usage", authMiddleware, readRateLimit);
app.use("/v1/billing/plan", authMiddleware, readRateLimit);
// Stripe webhook: IP-based, no auth (Stripe verifies via signature)
app.use("/v1/billing/webhook", webhookRateLimit);
// Templates: write-level limits (200 req/min)
app.use("/v1/templates/*", authMiddleware, writeRateLimit);
app.use("/v1/templates", authMiddleware, writeRateLimit);
// Voice: write-level limits (200 req/min)
app.use("/v1/voice/*", authMiddleware, writeRateLimit);

// Mount route handlers
app.route("/v1/messages", messages);
app.route("/v1/domains", domains);
app.route("/v1/webhooks", webhooks);
app.route("/v1/analytics", analytics);
app.route("/v1/suppressions", suppressions);
app.route("/v1/api-keys", apiKeysRouter);
app.route("/v1/account", account);
app.route("/v1/billing", billing);
app.route("/v1/templates", templatesRouter);
app.route("/v1/voice", voice);

// ─── 404 handler ────────────────────────────────────────────────────────────

app.notFound((c) => {
  return c.json(
    {
      error: {
        type: "not_found",
        message: `Route ${c.req.method} ${c.req.path} not found`,
        code: "route_not_found",
      },
    },
    404,
  );
});

// ─── Global error handler ──────────────────────────────────────────────────

app.onError((err, c) => {
  const reqId = c.get("requestId") ?? "unknown";

  console.error(`[${reqId}] Unhandled error:`, err);

  const isProduction = process.env.NODE_ENV === "production";

  return c.json(
    {
      error: {
        type: "server_error",
        message: isProduction
          ? "An internal server error occurred"
          : err.message,
        code: "internal_error",
        ...(isProduction ? {} : { stack: err.stack }),
      },
    },
    500,
  );
});

// ─── Server startup ─────────────────────────────────────────────────────────

const port = parseInt(process.env["PORT"] ?? "3001", 10);

console.log("=".repeat(60));
console.log("  Emailed API — Starting");
console.log(`  Port: ${port}`);
console.log(`  Environment: ${process.env.NODE_ENV ?? "development"}`);
console.log("=".repeat(60));

// Initialize OpenTelemetry
initTelemetry("emailed-api").catch((err) => {
  console.warn("[api] OpenTelemetry init failed:", err);
});

// Initialize Meilisearch index on startup (non-blocking)
initSearchIndex().catch((err) => {
  console.warn("[api] Meilisearch init failed (search will be unavailable):", err);
});

// Start the webhook delivery worker (BullMQ consumer)
startWebhookWorker();

// ─── Graceful shutdown ──────────────────────────────────────────────────────

let isShuttingDown = false;

async function shutdown(signal: string): Promise<void> {
  if (isShuttingDown) return;
  isShuttingDown = true;

  console.log(`\n[api] Received ${signal} — shutting down...`);

  const timeout = setTimeout(() => {
    console.error("[api] Shutdown timed out — forcing exit");
    process.exit(1);
  }, 15_000);

  try {
    // Close the webhook delivery worker
    await stopWebhookWorker();
    console.log("[api] Webhook worker stopped");

    // Close the BullMQ send queue
    await closeSendQueue();
    console.log("[api] Send queue closed");

    // Flush telemetry
    await shutdownTelemetry();
    console.log("[api] Telemetry shut down");

    // Close rate-limit Redis connection
    await closeRateLimitRedis();
    console.log("[api] Rate-limit Redis closed");

    // Close the database connection pool
    await closeConnection();
    console.log("[api] Database connections closed");

    clearTimeout(timeout);
    console.log("[api] Graceful shutdown complete");
    process.exit(0);
  } catch (error) {
    console.error("[api] Shutdown error:", error);
    clearTimeout(timeout);
    process.exit(1);
  }
}

process.on("SIGTERM", () => shutdown("SIGTERM"));
process.on("SIGINT", () => shutdown("SIGINT"));

// ─── Export for Bun serve() ─────────────────────────────────────────────────

export default {
  port,
  fetch: app.fetch,
};

export { app };