# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Emailed API
  description: |
    The Emailed API provides programmatic access to the Emailed email
    infrastructure platform. Send transactional and bulk emails, manage
    domains, configure webhooks, and query analytics.
  version: 0.1.0
  contact:
    name: Emailed Support
    url: https://emailed.dev/support
    email: support@emailed.dev
  license:
    name: Proprietary
    url: https://emailed.dev/terms
servers:
  - url: https://api.emailed.dev
    description: Production
  - url: https://api.staging.emailed.dev
    description: Staging
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /v1/messages:
    post:
      operationId: sendMessage
      summary: Send an email
      description: |
        Enqueues an email for delivery. Returns immediately with a message ID
        and a `202 Accepted` status. Use the message ID to poll for delivery
        status or configure webhooks for async notifications.
      tags: [Messages]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageInput"
      responses:
        "202":
          description: Message accepted for delivery
          headers:
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/json:
              schema:
                type: object
                required: [id, status, createdAt]
                properties:
                  id:
                    type: string
                    description: Unique message identifier
                    example: a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4
                  status:
                    type: string
                    enum: [queued, sending]
                    example: sending
                  createdAt:
                    type: string
                    format: date-time
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      operationId: listMessages
      summary: List messages
      description: |
        Returns a paginated list of messages. Supports cursor-based pagination
        and filtering by status or tag.
      tags: [Messages]
      parameters:
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: status
          in: query
          schema:
            type: string
            enum: [queued, sending, delivered, bounced, deferred, complained, failed]
        - name: tag
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated list of messages
          headers:
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
          content:
            application/json:
              schema:
                type: object
                required: [data, hasMore]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/MessageRecord"
                  cursor:
                    type: string
                    nullable: true
                    description: Cursor for the next page, or null if no more results
                  hasMore:
                    type: boolean
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/domains:
    post:
      operationId: addDomain
      summary: Add a domain
      description: |
        Register a new sending domain. Returns the required DNS records that
        must be configured before verification.
      tags: [Domains]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddDomainInput"
      responses:
        "201":
          description: Domain added successfully
          content:
            application/json:
              schema:
                type: object
                required: [data, message]
                properties:
                  data:
                    $ref: "#/components/schemas/DomainRecord"
                  message:
                    type: string
                    example: "Domain added. Configure the DNS records below, then verify."
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Domain already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/messages/search:
    get:
      operationId: searchMessages
      summary: Full-text email search
      description: Search messages via Meilisearch full-text index.
      tags: [Messages]
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: mailbox
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/LimitParam"
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/MessageRecord"
                  totalHits:
                    type: integer
                  processingTimeMs:
                    type: number
                  query:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          description: Search service unavailable
  /v1/templates:
    post:
      operationId: createTemplate
      summary: Create an email template
      tags: [Templates]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, subject]
              properties:
                name:
                  type: string
                subject:
                  type: string
                htmlBody:
                  type: string
                textBody:
                  type: string
                metadata:
                  type: object
      responses:
        "201":
          description: Template created
        "401":
          $ref: "#/components/responses/Unauthorized"
    get:
      operationId: listTemplates
      summary: List templates
      tags: [Templates]
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/CursorParam"
        - name: name
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated list of templates
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/suppressions:
    post:
      operationId: addSuppression
      summary: Add email to suppression list
      tags: [Suppressions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, domainId, reason]
              properties:
                email:
                  type: string
                  format: email
                domainId:
                  type: string
                reason:
                  type: string
                  enum: [bounce, complaint, unsubscribe, manual]
      responses:
        "201":
          description: Suppression added
        "401":
          $ref: "#/components/responses/Unauthorized"
    get:
      operationId: listSuppressions
      summary: List suppressions
      tags: [Suppressions]
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/CursorParam"
        - name: domain
          in: query
          schema:
            type: string
        - name: reason
          in: query
          schema:
            type: string
            enum: [bounce, complaint, unsubscribe, manual]
      responses:
        "200":
          description: Paginated suppression list
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/account:
    get:
      operationId: getAccount
      summary: Get account details
      tags: [Account]
      responses:
        "200":
          description: Account profile
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      operationId: updateAccount
      summary: Update account settings
      tags: [Account]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                billingEmail:
                  type: string
      responses:
        "200":
          description: Account updated
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/api-keys:
    post:
      operationId: createApiKey
      summary: Create an API key
      tags: [API Keys]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, scopes]
              properties:
                name:
                  type: string
                scopes:
                  type: array
                  items:
                    type: string
      responses:
        "201":
          description: API key created (key shown only once)
        "401":
          $ref: "#/components/responses/Unauthorized"
    get:
      operationId: listApiKeys
      summary: List API keys
      tags: [API Keys]
      responses:
        "200":
          description: List of API keys (secrets redacted)
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/billing/checkout:
    post:
      operationId: createCheckout
      summary: Create Stripe checkout session
      tags: [Billing]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [planId, successUrl, cancelUrl]
              properties:
                planId:
                  type: string
                  enum: [starter, professional, enterprise]
                successUrl:
                  type: string
                  format: uri
                cancelUrl:
                  type: string
                  format: uri
      responses:
        "200":
          description: Checkout session URL
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/billing/portal:
    post:
      operationId: createBillingPortal
      summary: Create Stripe billing portal session
      tags: [Billing]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [returnUrl]
              properties:
                returnUrl:
                  type: string
                  format: uri
      responses:
        "200":
          description: Portal session URL
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/billing/usage:
    get:
      operationId: getUsage
      summary: Get current billing period usage
      tags: [Billing]
      responses:
        "200":
          description: Usage statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      emailsSent:
                        type: integer
                      emailsLimit:
                        type: integer
                      percentUsed:
                        type: number
                      planTier:
                        type: string
                      limitExceeded:
                        type: boolean
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/billing/plan:
    get:
      operationId: getPlan
      summary: Get current plan details
      tags: [Billing]
      responses:
        "200":
          description: Current plan with limits
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/webhooks:
    post:
      operationId: createWebhook
      summary: Register a webhook
      description: |
        Create a webhook endpoint to receive real-time event notifications.
        The platform signs payloads with HMAC-SHA256 using the provided secret.
      tags: [Webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWebhookInput"
      responses:
        "201":
          description: Webhook created
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/WebhookRecord"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      operationId: listWebhooks
      summary: List webhooks
      tags: [Webhooks]
      responses:
        "200":
          description: List of webhooks
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        Pass a Bearer token in the Authorization header.
        Example: `Authorization: Bearer em_live_abc123...`
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Pass an API key in the X-API-Key header.
        Example: `X-API-Key: em_live_abc123...`
  headers:
    X-Request-Id:
      description: Unique identifier for the request, useful for support and debugging
      schema:
        type: string
        example: req_a1b2c3d4
    X-RateLimit-Limit:
      description: Maximum number of requests allowed in the current window
      schema:
        type: integer
        example: 1000
    X-RateLimit-Remaining:
      description: Number of requests remaining in the current window
      schema:
        type: integer
        example: 997
    X-RateLimit-Reset:
      description: Unix timestamp (seconds) when the rate limit window resets
      schema:
        type: integer
        example: 1712150400
  parameters:
    CursorParam:
      name: cursor
      in: query
      description: Pagination cursor from a previous response
      schema:
        type: string
    LimitParam:
      name: limit
      in: query
      description: Maximum number of items to return (1-100, default 20)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
    SendMessageInput:
      type: object
      required: [from, to, subject]
      properties:
        from:
          type: string
          format: email
          description: Sender email address
          example: noreply@example.com
        to:
          type: array
          items:
            type: string
            format: email
          minItems: 1
          maxItems: 50
          description: Recipient email addresses
          example: ["user@example.com"]
        cc:
          type: array
          items:
            type: string
            format: email
          description: CC recipients
        bcc:
          type: array
          items:
            type: string
            format: email
          description: BCC recipients
        subject:
          type: string
          maxLength: 998
          description: Email subject line
          example: Welcome to Emailed
        textBody:
          type: string
          description: Plain text body
        htmlBody:
          type: string
          description: HTML body
        tags:
          type: array
          items:
            type: string
          description: Tags for categorization and filtering
        metadata:
          type: object
          additionalProperties:
            type: string
          description: Custom key-value metadata (max 20 entries)
        scheduledAt:
          type: string
          format: date-time
          description: Schedule delivery for a future time
    MessageRecord:
      type: object
      required: [id, from, to, subject, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        from:
          type: string
          format: email
        to:
          type: array
          items:
            type: string
            format: email
        cc:
          type: array
          items:
            type: string
            format: email
        bcc:
          type: array
          items:
            type: string
            format: email
        subject:
          type: string
        status:
          type: string
          enum: [queued, sending, delivered, bounced, deferred, complained, failed]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deliveredAt:
          type: string
          format: date-time
          nullable: true
    AddDomainInput:
      type: object
      required: [domain]
      properties:
        domain:
          type: string
          description: The domain name to register
          example: example.com
    DomainRecord:
      type: object
      required: [id, domain, status, dnsRecords, spfVerified, dkimVerified, dmarcVerified, mxVerified, createdAt]
      properties:
        id:
          type: string
        domain:
          type: string
        status:
          type: string
          enum: [pending, verified, failed]
        dnsRecords:
          type: array
          items:
            $ref: "#/components/schemas/DnsRecord"
        spfVerified:
          type: boolean
        dkimVerified:
          type: boolean
        dmarcVerified:
          type: boolean
        mxVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
    DnsRecord:
      type: object
      required: [type, host, value]
      properties:
        type:
          type: string
          enum: [TXT, CNAME, MX, A, AAAA]
        host:
          type: string
        value:
          type: string
        priority:
          type: integer
          description: MX record priority
    CreateWebhookInput:
      type: object
      required: [url, events]
      properties:
        url:
          type: string
          format: uri
          description: The HTTPS URL to receive webhook events
          example: https://example.com/webhooks/emailed
        events:
          type: array
          items:
            type: string
            enum:
              - message.sent
              - message.delivered
              - message.bounced
              - message.deferred
              - message.dropped
              - message.complained
              - message.opened
              - message.clicked
              - domain.verified
              - domain.failed
              - contact.subscribed
              - contact.unsubscribed
          minItems: 1
          description: Event types to subscribe to
        secret:
          type: string
          description: HMAC signing secret for payload verification
        description:
          type: string
          maxLength: 256
        active:
          type: boolean
          default: true
    WebhookRecord:
      type: object
      required: [id, url, events, active, createdAt, updatedAt]
      properties:
        id:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
        secret:
          type: string
          description: Redacted in responses
        description:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [type, message, code]
          properties:
            type:
              type: string
              description: Error category
              example: validation_error
            message:
              type: string
              description: Human-readable error message
              example: "Invalid email address in 'to' field"
            code:
              type: string
              description: Machine-readable error code
              example: invalid_recipient
            context:
              type: object
              additionalProperties: true
              description: Additional error context
  responses:
    BadRequest:
      description: The request body or parameters are malformed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication credentials are missing or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated but insufficient permissions for this operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ValidationError:
      description: The request failed validation (invalid email, domain, etc.)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
        X-RateLimit-Limit:
          $ref: "#/components/headers/X-RateLimit-Limit"
        X-RateLimit-Remaining:
          $ref: "#/components/headers/X-RateLimit-Remaining"
        X-RateLimit-Reset:
          $ref: "#/components/headers/X-RateLimit-Reset"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: An unexpected server error occurred
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"